% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestage-core.R
\name{nestage_compare}
\alias{nestage_compare}
\title{Compute deltas and fold changes across models}
\usage{
nestage_compare(df)
}
\arguments{
\item{df}{Output from nestage_run_all() or rbind of many populations.}
}
\value{
Data.frame with comparisons per population.
}
\description{
Compute deltas and fold changes across models
}
