% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestage-core.R
\name{nestage_exact}
\alias{nestage_exact}
\title{Observed (Exact) estimator: Ny/N, Ne/N, and lambda}
\usage{
nestage_exact(A_obs, meta = list())
}
\arguments{
\item{A_obs}{Observed projection matrix.}

\item{meta}{Optional list with identifiers (for example, list(population = "Miz")).}
}
\value{
A list of class 'nestage_result' with fields: model, NyN, NeN, lambda, L, version, meta.
}
\description{
Observed (Exact) estimator: Ny/N, Ne/N, and lambda
}
