% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestage-core.R
\name{nestage_expected}
\alias{nestage_expected}
\title{Expected estimator: Ny/N, Ne/N, and lambda}
\usage{
nestage_expected(A_exp, meta = list())
}
\arguments{
\item{A_exp}{Expected projection matrix.}

\item{meta}{Optional list with identifiers.}
}
\value{
A list of class 'nestage_result'.
}
\description{
Expected estimator: Ny/N, Ne/N, and lambda
}
