% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestage-plots.R
\name{nestage_plot_generation_time}
\alias{nestage_plot_generation_time}
\title{Plot generation time (L) for Sex-only models (points only)}
\usage{
nestage_plot_generation_time(df)
}
\arguments{
\item{df}{Tidy table from nestage_run_all() or combined tables.}
}
\description{
Plot generation time (L) for Sex-only models (points only)
}
