% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestage-plots.R
\name{nestage_plot_ratios}
\alias{nestage_plot_ratios}
\title{Plot Ny/N and Ne/N by population and model (points only)}
\usage{
nestage_plot_ratios(df)
}
\arguments{
\item{df}{Tidy table from nestage_run_all() or combined tables.}
}
\description{
Plot Ny/N and Ne/N by population and model (points only)
}
